feat(vitepress): add articles component and related utils
Create a new articles component along with utility functions to support
it. This component will be used to display articles in a grid format,
utilizing vue's dynamic class binding for different grid layouts based
on the number of articles. Also included are improvements to the link
component and shared utility functions to enhance the overall user
experience and consistency of the site.